refactor: make dirs for header and landing
